+++ RenderEngines/ModalRenderEngine.cs
@@ -97,7 +97,7 @@ namespace RhinoCyclesCore.RenderEngines
 			//ViewCrc = e.Crc;
 		}
 
-		private uint maxSamples;
+		private int maxSamples;
 
 		/// <summary>
 		/// Entry point for a new render process. This is to be done in a separate thread.
@@ -112,7 +112,7 @@ namespace RhinoCyclesCore.RenderEngines
 			if (rw == null) return; // we don't have a window to write to...
 
 			var size = cyclesEngine.RenderDimension;
-			_samples = (uint)RcCore.It.EngineSettings.Samples;
+			_samples = RcCore.It.EngineSettings.Samples;
 			maxSamples = _samples;
 			cyclesEngine.TriggerCurrentViewportSettingsRequested();
 
@@ -131,7 +131,7 @@ namespace RhinoCyclesCore.RenderEngines
 			var sessionParams = new SessionParameters(client, renderDevice)
 			{
 				Experimental = false,
-				Samples = (int)_samples,
+				Samples = _samples,
 				TileSize = renderDevice.IsCpu ? new Size(32, 32) : new Size(RcCore.It.EngineSettings.TileX, RcCore.It.EngineSettings.TileY),
 				TileOrder = TileOrder.Center,
 				Threads = (uint)(renderDevice.IsGpu ? 0 : RcCore.It.EngineSettings.Threads),
@@ -162,7 +162,7 @@ namespace RhinoCyclesCore.RenderEngines
 			cyclesEngine.Session.PrepareRun();
 
 			// lets first reset session
-			cyclesEngine.Session.Reset((uint)size.Width, (uint)size.Height, _samples);
+			cyclesEngine.Session.Reset(size.Width, size.Height, _samples);
 			// then reset scene
 			cyclesEngine.Session.Scene.Reset();
 			// and actually start
